черновая верстка дополнения вывода о погоде

+++ app/src/main/java/com/example/weather/data/network/data/PostMain.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 public class PostMain {
     private float temp;
     /*давление*/
-    private int pressure;
+    private float pressure;
     /*влажность*/
     private int humidity;
 }
